modes.go: move out rate limiters to separate package
